requirements.txt
sounddevice
pydub
openai-whisper
ffmpeg-python
librosa
groq
python-dotenv
fastapi 
uvicorn 
python-multipart 
moto 
boto3 
aiohttp 
numpy 
wave 